+++ pypit/arload.py
@@ -96,10 +96,10 @@ def load_headers(datlines):
         skip = False
         for ch in chks:
             tfrhd = int(ch.split('.')[0])-1
-            kchk  = '.'.join(ch.split('.')[1:])
-            frhd  = whddict['{0:02d}'.format(tfrhd)]
-            # Perform test
-            if settings.spect['check'][ch] != str(headarr[frhd][kchk]).strip():
+            kchk = '.'.join(ch.split('.')[1:])
+            frhd = whddict['{0:02d}'.format(tfrhd)]
+            # JFH changed to in instead of !=
+            if ((settings.spect['check'][ch] in str(headarr[frhd][kchk]).strip()) == False):
                 print(ch, frhd, kchk)
                 print(settings.spect['check'][ch], str(headarr[frhd][kchk]).strip())
                 msgs.warn("The following file:"+msgs.newline()+datlines[i]+msgs.newline()+"is not taken with the settings.{0:s} detector".format(settings.argflag['run']['spectrograph'])+msgs.newline()+"Remove this file, or specify a different settings file.")
@@ -239,6 +239,8 @@ def load_frames(fitsdict, ind, det, frametype='<None>', msbias=None, trim=True):
         # Instrument specific read
         if settings.argflag['run']['spectrograph'] in ['keck_lris_blue', 'keck_lris_red']:
             temp, head0, _ = arlris.read_lris(fitsdict['directory'][ind[i]]+fitsdict['filename'][ind[i]], det=det)
+        elif settings.argflag['run']['spectrograph'] in ['keck_deimos']:
+            temp, head0, _ = ardeimos.read_deimos(fitsdict['directory'][ind[i]] + fitsdict['filename'][ind[i]])
         else:
             hdulist = fits.open(fitsdict['directory'][ind[i]]+fitsdict['filename'][ind[i]])
             temp = hdulist[settings.spect[dnum]['dataext01']].data