+++ pypit/arsciexp.py
@@ -740,7 +740,7 @@ class ScienceExposure:
                 # Use this detector? Need to check this after setting RA/DEC above
                 if settings.argflag['reduce']['detnum'] is not None:
                     msgs.warn("If your standard wasnt on this detector, you will have trouble..")
-                    if det != settings.argflag['reduce']['detnum']:
+                    if det not in map(int, settings.argflag['reduce']['detnum']):
                         continue
                 if settings.spect["mosaic"]["reduction"] == "ARMLSD":
                     arproc.reduce_multislit(self, sciframe, ind[0], fitsdict, det, standard=True)