+++ pypit/artrace.py
@@ -41,8 +41,7 @@ except NameError:
 
 
 def assign_slits(binarr, edgearr, ednum=100000, lor=-1):
-    """
-    This routine will traces the locations of the slit edges
+    """This routine will trace the locations of the slit edges
 
     Parameters
     ----------
@@ -382,6 +381,7 @@ def trace_objbg_image(slf, det, sciframe, slitn, objreg, bgreg, trim=2, triml=No
     msgs.info("Creating an image weighted by object pixels")
     rec_obj_img = np.zeros((sciframe.shape[0], sciframe.shape[1], nobj))
     for o in range(nobj):
+        #msgs.info("obj {:d}".format(o))
         obj = np.zeros(npix)
         obj[objreg[0][o]:objreg[1][o]+1] = 1
         scitmp = np.append(0.0, np.append(obj, 0.0))
@@ -405,6 +405,7 @@ def trace_objbg_image(slf, det, sciframe, slitn, objreg, bgreg, trim=2, triml=No
     msgs.info("Creating an image weighted by background pixels")
     rec_bg_img = np.zeros((sciframe.shape[0], sciframe.shape[1], nobj))
     for o in range(nobj):
+        #msgs.info("bkg {:d}".format(o))
         backtmp = np.append(0.0, np.append(bgreg[0][:, o] + bgreg[1][:, o], 0.0))
         bckframe = backtmp.reshape(1, -1).repeat(sciframe.shape[0], axis=0)
         bckspl = interpolate.RectBivariateSpline(xint, yint, bckframe,
@@ -853,7 +854,7 @@ def obj_trace_qa(slf, frame, ltrace, rtrace, objids, det,
     fig = plt.figure(dpi=1200)
 
     plt.rcParams['font.family'] = 'times new roman'
-    ticks_font = font_manager.FontProperties(family='times new roman', 
+    ticks_font = font_manager.FontProperties(family='times new roman',
        style='normal', size=16, weight='normal', stretch='normal')
     ax = plt.gca()
     for label in ax.get_yticklabels() :
@@ -925,7 +926,7 @@ def plot_orderfits(slf, model, ydata, xdata=None, xmodl=None, textplt="Slit",
 
     plt.rcdefaults()
     plt.rcParams['font.family']= 'times new roman'
-    
+
     # Outfil
     method = inspect.stack()[0][3]
     if 'Arc' in desc:
@@ -1636,7 +1637,7 @@ def trace_slits(slf, mstrace, det, pcadesc="", maskBadRows=False, min_sqm=30.):
 
     msgs.info("Preparing trace frame for slit detection")
     # Generate a binned (or smoothed) version of the trace frame
-    binarr = ndimage.uniform_filter(mstrace, size=(3, 1))
+    binarr = ndimage.uniform_filter(mstrace, size=(3, 1),mode='mirror')
     binbpx = slf._bpix[det-1].copy()
     plxbin = slf._pixlocn[det-1][:, :, 0].copy()
     plybin = slf._pixlocn[det-1][:, :, 1].copy()
@@ -1685,7 +1686,7 @@ def trace_slits(slf, mstrace, det, pcadesc="", maskBadRows=False, min_sqm=30.):
         sqmstrace[(sqmstrace < 1.0) & (sqmstrace >= 0.0)] = 1.0
         sqmstrace[(sqmstrace > -1.0) & (sqmstrace <= 0.0)] = -1.0
         # Apply a Sobel filter
-        filt = ndimage.sobel(sqmstrace, axis=1, mode='nearest')
+        filt = ndimage.sobel(sqmstrace, axis=1, mode=settings.argflag['trace']['slits']['sobel']['mode'])
         msgs.info("Applying bad pixel mask")
         filt *= (1.0 - binbpx)  # Apply to the bad pixel mask
         siglev = np.sign(filt)*(filt**2)/np.maximum(sqmstrace, min_sqm)
@@ -1694,14 +1695,16 @@ def trace_slits(slf, mstrace, det, pcadesc="", maskBadRows=False, min_sqm=30.):
         wr = np.where(siglev < -settings.argflag['trace']['slits']['sigdetect'])  # A negative gradient is a right edge
         tedges[wl] = -1.0
         tedges[wr] = +1.0
-        hdu = fits.PrimaryHDU(filt)
-        hdu.writeto("filt_{0:02d}.fits".format(det), overwrite=True)
-        hdu = fits.PrimaryHDU(sqmstrace)
-        hdu.writeto("sqmstrace_{0:02d}.fits".format(det), overwrite=True)
-        hdu = fits.PrimaryHDU(binarr)
-        hdu.writeto("binarr_{0:02d}.fits".format(det), overwrite=True)
-        hdu = fits.PrimaryHDU(siglev)
-        hdu.writeto("siglev_{0:02d}.fits".format(det), overwrite=True)
+        if False:
+            import astropy.io.fits as pyfits
+            hdu = pyfits.PrimaryHDU(filt)
+            hdu.writeto("filt_{0:02d}.fits".format(det), overwrite=True)
+            hdu = pyfits.PrimaryHDU(sqmstrace)
+            hdu.writeto("sqmstrace_{0:02d}.fits".format(det), overwrite=True)
+            hdu = pyfits.PrimaryHDU(binarr)
+            hdu.writeto("binarr_{0:02d}.fits".format(det), overwrite=True)
+            hdu = pyfits.PrimaryHDU(siglev)
+            hdu.writeto("siglev_{0:02d}.fits".format(det), overwrite=True)
         # Clean the edges
         wcl = np.where((ndimage.maximum_filter1d(siglev, 10, axis=1) == siglev) & (tedges == -1))
         wcr = np.where((ndimage.minimum_filter1d(siglev, 10, axis=1) == siglev) & (tedges == +1))
@@ -1748,8 +1751,6 @@ def trace_slits(slf, mstrace, det, pcadesc="", maskBadRows=False, min_sqm=30.):
             edgearr = np.copy(nedgear)
 
     # Assign a number to each of the edges
-    msgs.info("Matching slit edges")
-
 #    _edgearr = edgearr.copy()
 #    t = time.clock()
 #    _lcnt, _rcnt = arcytrace.match_edges(_edgearr, ednum)
@@ -1983,7 +1984,6 @@ def trace_slits(slf, mstrace, det, pcadesc="", maskBadRows=False, min_sqm=30.):
         #msgs.info("Ignoring any slit that spans < {0:3.2f}x{1:d} pixels on the detector".format(settings.argflag['trace']['slits']['fracignore'], int(edgearr.shape[0]*binby)))
         msgs.info("Ignoring any slit that spans < {0:3.2f}x{1:d} pixels on the detector".format(settings.argflag['trace']['slits']['fracignore'], int(edgearr.shape[0])))
         fracpix = int(settings.argflag['trace']['slits']['fracignore']*edgearr.shape[0])
-
 #        print('calling ignore_orders')
 #        t = time.clock()
 #        _edgearr = edgearr.copy()
@@ -3469,7 +3469,6 @@ def multislit_tilt(slf, msarc, det, maskval=-999999.9):
                 debugger.set_trace()
         elif settings.argflag['trace']['slits']['tilts']['method'].lower() == "pca":
             tilts = polytilts.copy()
-
 
     # Now do the QA
     msgs.info("Preparing arc tilt QA data")